# documentation/v1_swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Notify ME",
    "description": "Api Documentation associated with Notify ME Application",
    "version": "1.0"
  },
  "produces": ["application/json"],
...
  "paths": {
    "/login": {
      "post": {
        "tags": ["Authentication"],
        "description": "Login Api Associated with Notify ME",
        "parameters": [{
          "name": "email",
          "required": true,
          "in": "formData",
          "type": "string"
        },
          {
            "name": "password",
            "required": true,
            "in": "formData",
            "type": "string"
          },
          {
            "name": "device_type",
            "required": true,
            "in": "formData",
            "type": "string"
          },
          {
            "name": "device_token",
            "required": true,
            "in": "formData",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/logout":{
      "post": {
        "tags": ["Authentication"],
        "description": "Logout Api Associated with Notify ME",
        "parameters": [
          {
            "name": "Authorization",
            "description" : "Token that has neared its expiry",
            "in": "header"
          },
          {
            "name": "device_token",
            "required": true,
            "in": "formData",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/refresh_token": {
      "get": {
        "tags": ["Authentication"],
        "description": "Refresh a Token that has neared Expiry",
        "parameters": [{
          "name": "Authorization",
          "description" : "Token that has neared its expiry",
          "in": "header"
        }],
        "responses": {}
      }
    },
    "/forgot_password": {
      "put": {
        "tags": ["Authentication"],
        "description": "Sends a Reset Password To Users",
        "parameters": [{
          "name": "email",
          "description" : "The Email Registered with The Application",
          "in": "formData"
        }],
        "responses": {}
      }
    },
...
  }
}